pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.3"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlmodel = "^0.0.22"
sqlalchemy = "^2.0.36"
alembic = "^1.13.3"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
asyncpg = "^0.30.0"
pyjwt = "^2.9.0"
bcrypt = "^4.2.0"
python-multipart = "^0.0.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
